Move sticker model into separate file
